requirements.txt
flask
flask-restful
flask-swagger-ui
passlib
PyJWT
marshmallow
pyyaml
pyOpenSSL
pytest
pytest-flask
pytest-cov
pytest-mock